requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==3.2.12
djangorestframework==3.14.0
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
kombu==5.3.2
packaging==23.2
phonenumbers==8.13.22
prompt-toolkit==3.0.39
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
vine==5.0.0
wcwidth==0.2.8
